x11: improve the vertical placement of rotated text (but it's still bad)
Rendering of rotated enhanced text in x11 is done by gplt_x11.c:DrawRotated.
Unfortunately the procedure used results in incorrect vertical placement of
rotated text.  The problem apparently dates back to the original introduction
of this routine in version 4.

I frankly don't understand the code in that routine, but the magnitude of
misplacement is clearly proportional to the sin of the the rotation angle.
In the absence of a correct fix, this commit adds an empirical correction
at the call site in x11.trm.   It does not scale very well with font size.

Bug #2661