Add a mechanism for failed evaluation to return NaN rather than int_error()
The new flag eval_fail_soft is similar to the flag string_result_only.
If an undefined variable or function is encountered during evaluation
the program normally calls int_error().  This prevents continuing with
normal sequential execution of commands in a script.
Setting eval_fail_soft = TRUE causes such a failure to instead return NaN.

The only user at this point is clone_linked_axes(), and it is applied
only to testing a named variable or function.
This allows the "set link" command to fail soft if the linking functions
use a variable or function that has not yet been defined.

Example:
    # foo is not (yet) defined
    set link x2 via foo*x*x inv sqrt(x/foo)

Bug #2677