Treat use of pseudofile '-' inside a multiplot as an error
1) "remultiplot" or "replot" or mousing in generate can now replay
the command sequence that created a multiplot.  If this includes
a plot command using pseudofile '-' as a data source, it would fail.
Worse, depending on the terminal type it might hang (x11, wxt, others?)
or even segfault (wxt but not 100% reproducible).
Related: Bug #2679

- catch this case in df_gets() and call
	int_error(NO_CARET, "cannot read data from '-' during remultiplot");

2) Fortunately it works to define and use a datablock instead.
This was working already, but $GPVAL_LAST_MULTIPLOT was left with an
fragmentary datablock definition line.  It was filtered out during
multiplot replay, but if save and reloaded this would be an error.

- do not pass datablock definition lines to append_multiplot_line()

3) The wxt terminal in particular behaved badly on the Bug #2679
reproducer.  Hitting ^C to regain control produced a spurious error
"input select error" and occasionally a segfault.

- the ^C is now caught cleanly in wxt_waitforinput() without calling
int_error().

I did not pin down the cause of the sporadic segfaults.