wxt: Exception handler for mouse event processing
Ever since the gtk version of wxWidgets moved from gtk2.8 to gtk3
the threading model of gnuplot's wxt terminal has been inconsistent
with the implementation in the wx + gtk libraries.  This made
gnuplot error handling unreliable if the wxt terminal was active.

The problem in particular manifests when int_error() is called from
inside an action triggered by a mouse event.
Mouse events include pan/zoom, 3D rotation, and hot key bindings.
Actions are gnuplot commands executed by the routine do_event().
Gnuplot's int_error() prints an error message and then returns to top
level command line processing via a LONGJMP in bail_to_command_line().
But the wx/gtk library state at the time the wxt terminal calls back
into the core routines via do_event() is not clean with respect to the
state originally saved for LONGJMP to return to.  Thus if the operations
requested by do_event() cause a core routine to exit via int_error(),
a segfault from somewhere in wxt is likely to happen soon after.

The best fix for this would probably be to refactor the wxt terminal
into inboard/outboard halves with the display and mousing handled by
a separate process.  That is the way the x11 and qt terminals work.
But that is a project for another day.

What this commit does instead is add an exception handler wrapping
the call to do_event() in the wxt terminal event processing code.
The core routine bail_to_command_line() can then LONGJMP to this
handler rather than directly back to the gnuplot command line.
Hopefully (because I am not entirely sure this is guaranteed) the
wxt event handling then finishes cleanly and we end up back at
a command line anyway.

Tested only under linux.
I am not confident that this is a complete solution, but it is only
invoked if the program was highly likely to segfault anyhow.

Bugs #2680 #2683