Draw a single dot for arrows that are too short for a vector
Gnuplot version 5.2 always drew at least a dot, possibly a single
pixel, for every arrow shaft.  The corresponding arrowheads were
not good.   Version 5.4 revised this to [optionally] scale the
arrowheads and orient them correctly, but if the arrow shaft would
be a single pixel or less it was not draw at all and neither was
the corresponding arrowhead.

This returns to the 5.2 behavior of always drawing at least a dot
at the position of the head of the vector.

Bug #2690